对用户、部门的操作按钮添加权限控制，对审核流程页面添加超级管理权限控制

+++ src/main/java/com/meiren/web/acl/UserModule.java
@@ -65,6 +65,11 @@ public class UserModule extends BaseController {
     public ApiResult resign(HttpServletRequest request, HttpServletResponse response) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long userId = RequestUtil.getLong(request, "id");
             if (userId == null) {
                 result.setError("id not null");
@@ -102,6 +107,11 @@ public class UserModule extends BaseController {
                                  HttpServletResponse response) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long userId = RequestUtil.getLong(request, "userId");
             if (userId == null) {
                 result.setError("id not null");
@@ -172,6 +182,11 @@ public class UserModule extends BaseController {
                                       HttpServletResponse response, @PathVariable String type) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long initId = RequestUtil.getLong(request, "dataId");
             Long selectedId = RequestUtil.getLong(request, "selectedId");
             Long uid = RequestUtil.getLong(request, "uid");
@@ -279,6 +294,11 @@ public class UserModule extends BaseController {
                                  HttpServletResponse response, @PathVariable String type) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long initId = RequestUtil.getLong(request, "dataId");
             Long selectedId = RequestUtil.getLong(request, "selectedId");
             Long uid = RequestUtil.getLong(request, "uid");
@@ -385,6 +405,11 @@ public class UserModule extends BaseController {
     public ApiResult hierarchySet(HttpServletRequest request, HttpServletResponse response, AclUserEntity aclUserEntity) {
         ApiResult result = new ApiResult();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long id = this.checkId(request);
             Map<String, Object> paramMap = new HashMap<>();
             paramMap.put("hierarchyId", aclUserEntity.getHierarchyId());
@@ -472,6 +497,11 @@ public class UserModule extends BaseController {
         ApiResult result = new ApiResult();
         Map<String, Object> delMap = new HashMap<>();
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             Long id = this.checkId(request);
             delMap.put("id", id);
             result = aclUserService.deleteAclUser(delMap);
@@ -495,6 +525,11 @@ public class UserModule extends BaseController {
         ApiResult result = new ApiResult();
         Map<String, Object> delMap = new HashMap<>();
         try {
+            AclUserEntity userEntity = this.getUser(request);
+            if(!this.hasUserAll(userEntity)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             UserStatusEnum userStatusEnum = !status.toLowerCase().equals("disable") ? UserStatusEnum.NORMAL : UserStatusEnum.DISABLE;
             Long id = this.checkId(request);
             delMap.put("id", id);
@@ -524,6 +559,11 @@ public class UserModule extends BaseController {
         String[] ids = request.getParameterValues("ids[]");
         List<String> idsList = Arrays.asList(ids);
         try {
+            AclUserEntity user = this.getUser(request);
+            if(!this.hasUserAll(user)){
+                result.setError("您没有权限操作用户！");
+                return result;
+            }
             delMap.put("inIds", idsList);
             result = aclUserService.deleteAclUser(delMap);
         } catch (Exception e) {
@@ -596,6 +636,7 @@ public class UserModule extends BaseController {
      * @param response
      * @return
      */
+    @AuthorityToken(needToken = {"meiren.acl.user.all"})
     @RequestMapping(value = "goTo/{type}", method = RequestMethod.GET)
     @ResponseBody
     public ModelAndView goTo(HttpServletRequest request, HttpServletResponse response, @PathVariable String type) {
